+++ crates/hyperswitch_domain_models/src/payment_methods.rs
@@ -11,15 +11,13 @@ use common_utils::crypto::OptionalEncryptableValue;
 use common_utils::{crypto::Encryptable, encryption::Encryption, types::keymanager::ToEncryptable};
 use common_utils::{
     errors::{CustomResult, ParsingError, ValidationError},
-    ext_traits::OptionExt,
-    id_type, pii, type_name,
-    types::{keymanager, CreatedBy},
+    id_type, pii,
+    types::CreatedBy,
 };
 pub use diesel_models::{
     enums as storage_enums, PaymentMethodUpdate as StoragePaymentMethodUpdate,
 };
-use error_stack::ResultExt;
-use hyperswitch_masking::{ExposeInterface, PeekInterface, Secret};
+use hyperswitch_masking::{ExposeInterface, Secret};
 #[cfg(feature = "v1")]
 use router_env::logger;
 #[cfg(feature = "v2")]
@@ -28,8 +26,6 @@ use rustc_hash::FxHashMap;
 use serde_json::Value;
 use time::PrimitiveDateTime;
 
-#[cfg(feature = "v1")]
-use crate::type_encryption::AsyncLift;
 #[cfg(feature = "v2")]
 use crate::{address::Address, platform::Initiator};
 use crate::{
@@ -37,7 +33,6 @@ use crate::{
     merchant_key_store::MerchantKeyStore,
     payment_method_data as domain_payment_method_data,
     transformers::ForeignTryFrom,
-    type_encryption::{crypto_operation, CryptoOperation},
 };
 
 #[derive(Debug, serde::Deserialize, serde::Serialize, Clone)]
@@ -415,510 +410,6 @@ impl
     }
 }
 
-#[cfg(feature = "v1")]
-#[async_trait::async_trait]
-impl super::behaviour::Conversion for PaymentMethod {
-    type DstType = diesel_models::payment_method::PaymentMethod;
-    type NewDstType = diesel_models::payment_method::PaymentMethodNew;
-    async fn convert(self) -> CustomResult<Self::DstType, ValidationError> {
-        let (vault_type, external_vault_source) = self.vault_source_details.into();
-        // Note: caller must ensure customer_id is not null before calling convert as storage model requires it.
-        Ok(Self::DstType {
-            customer_id: self.customer_id.get_required_value("customer_id")?,
-            merchant_id: self.merchant_id,
-            payment_method_id: self.payment_method_id,
-            accepted_currency: self.accepted_currency,
-            scheme: self.scheme,
-            token: self.token,
-            cardholder_name: self.cardholder_name,
-            issuer_name: self.issuer_name,
-            issuer_country: self.issuer_country,
-            payer_country: self.payer_country,
-            is_stored: self.is_stored,
-            swift_code: self.swift_code,
-            direct_debit_token: self.direct_debit_token,
-            created_at: self.created_at,
-            last_modified: self.last_modified,
-            payment_method: self.payment_method,
-            payment_method_type: self.payment_method_type,
-            payment_method_issuer: self.payment_method_issuer,
-            payment_method_issuer_code: self.payment_method_issuer_code,
-            metadata: self.metadata,
-            payment_method_data: self.payment_method_data.map(|val| val.into()),
-            locker_id: self.locker_id,
-            last_used_at: self.last_used_at,
-            connector_mandate_details: self.connector_mandate_details,
-            customer_acceptance: self.customer_acceptance,
-            status: self.status,
-            network_transaction_id: self.network_transaction_id,
-            client_secret: self.client_secret,
-            payment_method_billing_address: self
-                .payment_method_billing_address
-                .map(|val| val.into()),
-            updated_by: self.updated_by,
-            version: self.version,
-            network_token_requestor_reference_id: self.network_token_requestor_reference_id,
-            network_token_locker_id: self.network_token_locker_id,
-            network_token_payment_method_data: self
-                .network_token_payment_method_data
-                .map(|val| val.into()),
-            external_vault_source,
-            vault_type,
-            created_by: self.created_by.map(|created_by| created_by.to_string()),
-            last_modified_by: self
-                .last_modified_by
-                .map(|last_modified_by| last_modified_by.to_string()),
-            customer_details: self.customer_details.map(|val| val.into()),
-            locker_fingerprint_id: self.locker_fingerprint_id,
-            network_tokenization_data: self.network_tokenization_data.map(|val| val.into()),
-        })
-    }
-
-    async fn convert_back(
-        state: &keymanager::KeyManagerState,
-        item: Self::DstType,
-        key: &Secret<Vec<u8>>,
-        key_manager_identifier: keymanager::Identifier,
-    ) -> CustomResult<Self, ValidationError>
-    where
-        Self: Sized,
-    {
-        // Decrypt encrypted fields first
-        let (
-            payment_method_data,
-            payment_method_billing_address,
-            network_token_payment_method_data,
-            network_tokenization_data,
-            customer_details,
-        ) = async {
-            let payment_method_data = item
-                .payment_method_data
-                .async_lift(|inner| async {
-                    crypto_operation(
-                        state,
-                        type_name!(Self::DstType),
-                        CryptoOperation::DecryptOptional(inner),
-                        key_manager_identifier.clone(),
-                        key.peek(),
-                    )
-                    .await
-                    .and_then(|val| val.try_into_optionaloperation())
-                })
-                .await?;
-
-            let payment_method_billing_address = item
-                .payment_method_billing_address
-                .async_lift(|inner| async {
-                    crypto_operation(
-                        state,
-                        type_name!(Self::DstType),
-                        CryptoOperation::DecryptOptional(inner),
-                        key_manager_identifier.clone(),
-                        key.peek(),
-                    )
-                    .await
-                    .and_then(|val| val.try_into_optionaloperation())
-                })
-                .await?;
-
-            let network_token_payment_method_data = item
-                .network_token_payment_method_data
-                .async_lift(|inner| async {
-                    crypto_operation(
-                        state,
-                        type_name!(Self::DstType),
-                        CryptoOperation::DecryptOptional(inner),
-                        key_manager_identifier.clone(),
-                        key.peek(),
-                    )
-                    .await
-                    .and_then(|val| val.try_into_optionaloperation())
-                })
-                .await?;
-
-            let network_tokenization_data = item
-                .network_tokenization_data
-                .async_lift(|inner| async {
-                    crypto_operation(
-                        state,
-                        type_name!(Self::DstType),
-                        CryptoOperation::DecryptOptional(inner),
-                        key_manager_identifier.clone(),
-                        key.peek(),
-                    )
-                    .await
-                    .and_then(|val| val.try_into_optionaloperation())
-                })
-                .await?;
-
-            let customer_details = item
-                .customer_details
-                .async_lift(|inner| async {
-                    crypto_operation(
-                        state,
-                        type_name!(Self::DstType),
-                        CryptoOperation::DecryptOptional(inner),
-                        key_manager_identifier.clone(),
-                        key.peek(),
-                    )
-                    .await
-                    .and_then(|val| val.try_into_optionaloperation())
-                })
-                .await?;
-
-            Ok::<_, error_stack::Report<common_utils::errors::CryptoError>>((
-                payment_method_data,
-                payment_method_billing_address,
-                network_token_payment_method_data,
-                network_tokenization_data,
-                customer_details,
-            ))
-        }
-        .await
-        .change_context(ValidationError::InvalidValue {
-            message: "Failed while decrypting payment method data".to_string(),
-        })?;
-
-        let vault_source_details = PaymentMethodVaultSourceDetails::try_from((
-            item.vault_type,
-            item.external_vault_source,
-        ))?;
-
-        // Construct the domain type
-        // Storage always has customer_id, wrap in Some for domain
-        Ok(Self {
-            customer_id: Some(item.customer_id),
-            merchant_id: item.merchant_id,
-            payment_method_id: item.payment_method_id,
-            accepted_currency: item.accepted_currency,
-            scheme: item.scheme,
-            token: item.token,
-            cardholder_name: item.cardholder_name,
-            issuer_name: item.issuer_name,
-            issuer_country: item.issuer_country,
-            payer_country: item.payer_country,
-            is_stored: item.is_stored,
-            swift_code: item.swift_code,
-            direct_debit_token: item.direct_debit_token,
-            created_at: item.created_at,
-            last_modified: item.last_modified,
-            payment_method: item.payment_method,
-            payment_method_type: item.payment_method_type,
-            payment_method_issuer: item.payment_method_issuer,
-            payment_method_issuer_code: item.payment_method_issuer_code,
-            metadata: item.metadata,
-            payment_method_data,
-            locker_id: item.locker_id,
-            last_used_at: item.last_used_at,
-            connector_mandate_details: item.connector_mandate_details,
-            customer_acceptance: item.customer_acceptance,
-            status: item.status,
-            network_transaction_id: item.network_transaction_id,
-            client_secret: item.client_secret,
-            payment_method_billing_address,
-            updated_by: item.updated_by,
-            version: item.version,
-            network_token_requestor_reference_id: item.network_token_requestor_reference_id,
-            network_token_locker_id: item.network_token_locker_id,
-            network_token_payment_method_data,
-            vault_source_details,
-            created_by: item
-                .created_by
-                .and_then(|created_by| created_by.parse::<CreatedBy>().ok()),
-            last_modified_by: item
-                .last_modified_by
-                .and_then(|last_modified_by| last_modified_by.parse::<CreatedBy>().ok()),
-            customer_details,
-            locker_fingerprint_id: item.locker_fingerprint_id,
-            network_tokenization_data,
-            storage_type: None,
-        })
-    }
-
-    async fn construct_new(self) -> CustomResult<Self::NewDstType, ValidationError> {
-        let (vault_type, external_vault_source) = self.vault_source_details.into();
-        // Note: caller must ensure customer_id is not null before calling convert as storage model requires it.
-        Ok(Self::NewDstType {
-            customer_id: self.customer_id.get_required_value("customer_id")?,
-            merchant_id: self.merchant_id,
-            payment_method_id: self.payment_method_id,
-            accepted_currency: self.accepted_currency,
-            scheme: self.scheme,
-            token: self.token,
-            cardholder_name: self.cardholder_name,
-            issuer_name: self.issuer_name,
-            issuer_country: self.issuer_country,
-            payer_country: self.payer_country,
-            is_stored: self.is_stored,
-            swift_code: self.swift_code,
-            direct_debit_token: self.direct_debit_token,
-            created_at: self.created_at,
-            last_modified: self.last_modified,
-            payment_method: self.payment_method,
-            payment_method_type: self.payment_method_type,
-            payment_method_issuer: self.payment_method_issuer,
-            payment_method_issuer_code: self.payment_method_issuer_code,
-            metadata: self.metadata,
-            payment_method_data: self.payment_method_data.map(|val| val.into()),
-            locker_id: self.locker_id,
-            last_used_at: self.last_used_at,
-            connector_mandate_details: self.connector_mandate_details,
-            customer_acceptance: self.customer_acceptance,
-            status: self.status,
-            network_transaction_id: self.network_transaction_id,
-            client_secret: self.client_secret,
-            payment_method_billing_address: self
-                .payment_method_billing_address
-                .map(|val| val.into()),
-            updated_by: self.updated_by,
-            version: self.version,
-            network_token_requestor_reference_id: self.network_token_requestor_reference_id,
-            network_token_locker_id: self.network_token_locker_id,
-            network_token_payment_method_data: self
-                .network_token_payment_method_data
-                .map(|val| val.into()),
-            external_vault_source,
-            vault_type,
-            created_by: self.created_by.map(|created_by| created_by.to_string()),
-            last_modified_by: self
-                .last_modified_by
-                .map(|last_modified_by| last_modified_by.to_string()),
-            customer_details: self.customer_details.map(|val| val.into()),
-            locker_fingerprint_id: self.locker_fingerprint_id,
-            network_tokenization_data: self.network_tokenization_data.map(|val| val.into()),
-        })
-    }
-}
-
-#[cfg(feature = "v2")]
-#[async_trait::async_trait]
-impl super::behaviour::Conversion for PaymentMethod {
-    type DstType = diesel_models::payment_method::PaymentMethod;
-    type NewDstType = diesel_models::payment_method::PaymentMethodNew;
-    async fn convert(self) -> CustomResult<Self::DstType, ValidationError> {
-        Ok(Self::DstType {
-            customer_id: self.customer_id,
-            merchant_id: self.merchant_id,
-            id: self.id,
-            created_at: self.created_at,
-            last_modified: self.last_modified,
-            payment_method_type_v2: self.payment_method_type,
-            payment_method_subtype: self.payment_method_subtype,
-            payment_method_data: self.payment_method_data.map(|val| val.into()),
-            locker_id: self.locker_id.map(|id| id.get_string_repr().clone()),
-            last_used_at: self.last_used_at,
-            connector_mandate_details: self.connector_mandate_details.map(|cmd| cmd.into()),
-            customer_acceptance: self.customer_acceptance,
-            status: self.status,
-            network_transaction_id: self.network_transaction_id.map(Secret::new),
-            client_secret: self.client_secret,
-            payment_method_billing_address: self
-                .payment_method_billing_address
-                .map(|val| val.into()),
-            updated_by: self.updated_by,
-            locker_fingerprint_id: self.locker_fingerprint_id,
-            version: self.version,
-            network_token_requestor_reference_id: self.network_token_requestor_reference_id,
-            network_token_locker_id: self.network_token_locker_id,
-            network_token_payment_method_data: self
-                .network_token_payment_method_data
-                .map(|val| val.into()),
-            external_vault_source: self.external_vault_source,
-            external_vault_token_data: self.external_vault_token_data.map(|val| val.into()),
-            vault_type: self.vault_type,
-            created_by: self.created_by.map(|created_by| created_by.to_string()),
-            last_modified_by: self
-                .last_modified_by
-                .map(|last_modified_by| last_modified_by.to_string()),
-            customer_details: self.customer_details.map(|val| val.into()),
-            network_tokenization_data: self.network_tokenization_data.map(|val| val.into()),
-        })
-    }
-
-    async fn convert_back(
-        state: &keymanager::KeyManagerState,
-        storage_model: Self::DstType,
-        key: &Secret<Vec<u8>>,
-        key_manager_identifier: keymanager::Identifier,
-    ) -> CustomResult<Self, ValidationError>
-    where
-        Self: Sized,
-    {
-        use common_utils::ext_traits::ValueExt;
-
-        async {
-            let decrypted_data = crypto_operation(
-                state,
-                type_name!(Self::DstType),
-                CryptoOperation::BatchDecrypt(EncryptedPaymentMethod::to_encryptable(
-                    EncryptedPaymentMethod {
-                        payment_method_data: storage_model.payment_method_data,
-                        payment_method_billing_address: storage_model
-                            .payment_method_billing_address,
-                        network_token_payment_method_data: storage_model
-                            .network_token_payment_method_data,
-                        external_vault_token_data: storage_model.external_vault_token_data,
-                        customer_details: storage_model.customer_details,
-                        network_tokenization_data: storage_model.network_tokenization_data,
-                    },
-                )),
-                key_manager_identifier,
-                key.peek(),
-            )
-            .await
-            .and_then(|val| val.try_into_batchoperation())?;
-
-            let data = EncryptedPaymentMethod::from_encryptable(decrypted_data)
-                .change_context(common_utils::errors::CryptoError::DecodingFailed)
-                .attach_printable("Invalid batch operation data")?;
-
-            let payment_method_billing_address = data
-                .payment_method_billing_address
-                .map(|billing| {
-                    billing.deserialize_inner_value(|value| value.parse_value("Address"))
-                })
-                .transpose()
-                .change_context(common_utils::errors::CryptoError::DecodingFailed)
-                .attach_printable("Error while deserializing Address")?;
-
-            let payment_method_data = data
-                .payment_method_data
-                .map(|payment_method_data| {
-                    payment_method_data
-                        .deserialize_inner_value(|value| value.parse_value("Payment Method Data"))
-                })
-                .transpose()
-                .change_context(common_utils::errors::CryptoError::DecodingFailed)
-                .attach_printable("Error while deserializing Payment Method Data")?;
-
-            let network_token_payment_method_data = data
-                .network_token_payment_method_data
-                .map(|network_token_payment_method_data| {
-                    network_token_payment_method_data.deserialize_inner_value(|value| {
-                        value.parse_value("Network token Payment Method Data")
-                    })
-                })
-                .transpose()
-                .change_context(common_utils::errors::CryptoError::DecodingFailed)
-                .attach_printable("Error while deserializing Network token Payment Method Data")?;
-
-            let customer_details = data
-                .customer_details
-                .map(|customer_details| {
-                    customer_details.deserialize_inner_value(|value| {
-                        value.parse_value("Payment Method Customer Details")
-                    })
-                })
-                .transpose()
-                .change_context(common_utils::errors::CryptoError::DecodingFailed)
-                .attach_printable("Error while deserializing Payment Method Customer Details")?;
-
-            let external_vault_token_data = data
-                .external_vault_token_data
-                .map(|external_vault_token_data| {
-                    external_vault_token_data.deserialize_inner_value(|value| {
-                        value.parse_value("External Vault Token Data")
-                    })
-                })
-                .transpose()
-                .change_context(common_utils::errors::CryptoError::DecodingFailed)
-                .attach_printable("Error while deserializing External Vault Token Data")?;
-
-            let network_tokenization_data = data
-                .network_tokenization_data
-                .map(|tokenization_data| {
-                    tokenization_data.deserialize_inner_value(|value| {
-                        value.parse_value("Network Tokenization Data")
-                    })
-                })
-                .transpose()
-                .change_context(common_utils::errors::CryptoError::DecodingFailed)
-                .attach_printable("Error while deserializing Network Tokenization Data")?;
-
-            Ok::<Self, error_stack::Report<common_utils::errors::CryptoError>>(Self {
-                customer_id: storage_model.customer_id,
-                merchant_id: storage_model.merchant_id,
-                id: storage_model.id,
-                created_at: storage_model.created_at,
-                last_modified: storage_model.last_modified,
-                payment_method_type: storage_model.payment_method_type_v2,
-                payment_method_subtype: storage_model.payment_method_subtype,
-                payment_method_data,
-                locker_id: storage_model.locker_id.map(VaultId::generate),
-                last_used_at: storage_model.last_used_at,
-                connector_mandate_details: storage_model.connector_mandate_details.map(From::from),
-                customer_acceptance: storage_model.customer_acceptance,
-                status: storage_model.status,
-                network_transaction_id: storage_model
-                    .network_transaction_id
-                    .map(ExposeInterface::expose),
-                client_secret: storage_model.client_secret,
-                payment_method_billing_address,
-                updated_by: storage_model.updated_by,
-                locker_fingerprint_id: storage_model.locker_fingerprint_id,
-                version: storage_model.version,
-                network_token_requestor_reference_id: storage_model
-                    .network_token_requestor_reference_id,
-                network_token_locker_id: storage_model.network_token_locker_id,
-                network_token_payment_method_data,
-                external_vault_source: storage_model.external_vault_source,
-                external_vault_token_data,
-                vault_type: storage_model.vault_type,
-                created_by: storage_model
-                    .created_by
-                    .and_then(|created_by| created_by.parse::<CreatedBy>().ok()),
-                last_modified_by: storage_model
-                    .last_modified_by
-                    .and_then(|last_modified_by| last_modified_by.parse::<CreatedBy>().ok()),
-                customer_details,
-                network_tokenization_data,
-            })
-        }
-        .await
-        .change_context(ValidationError::InvalidValue {
-            message: "Failed while decrypting payment method data".to_string(),
-        })
-    }
-
-    async fn construct_new(self) -> CustomResult<Self::NewDstType, ValidationError> {
-        Ok(Self::NewDstType {
-            customer_id: self.customer_id,
-            merchant_id: self.merchant_id,
-            id: self.id,
-            created_at: self.created_at,
-            last_modified: self.last_modified,
-            payment_method_type_v2: self.payment_method_type,
-            payment_method_subtype: self.payment_method_subtype,
-            payment_method_data: self.payment_method_data.map(|val| val.into()),
-            locker_id: self.locker_id.map(|id| id.get_string_repr().clone()),
-            last_used_at: self.last_used_at,
-            connector_mandate_details: self.connector_mandate_details.map(|cmd| cmd.into()),
-            customer_acceptance: self.customer_acceptance,
-            status: self.status,
-            network_transaction_id: self.network_transaction_id,
-            client_secret: self.client_secret,
-            payment_method_billing_address: self
-                .payment_method_billing_address
-                .map(|val| val.into()),
-            updated_by: self.updated_by,
-            locker_fingerprint_id: self.locker_fingerprint_id,
-            version: self.version,
-            network_token_requestor_reference_id: self.network_token_requestor_reference_id,
-            network_token_locker_id: self.network_token_locker_id,
-            network_token_payment_method_data: self
-                .network_token_payment_method_data
-                .map(|val| val.into()),
-            external_vault_token_data: self.external_vault_token_data.map(|val| val.into()),
-            vault_type: self.vault_type,
-            created_by: self.created_by.map(|created_by| created_by.to_string()),
-            last_modified_by: self
-                .last_modified_by
-                .map(|last_modified_by| last_modified_by.to_string()),
-            customer_details: self.customer_details.map(|val| val.into()),
-        })
-    }
-}
-
 #[cfg(feature = "v2")]
 #[derive(Clone, Debug, router_derive::ToEncryption)]
 pub struct PaymentMethodSession {
@@ -937,109 +428,6 @@ pub struct PaymentMethodSession {
     pub associated_token_id: Option<id_type::GlobalTokenId>,
     pub storage_type: common_enums::StorageType,
     pub keep_alive: bool,
-}
-
-#[cfg(feature = "v2")]
-#[async_trait::async_trait]
-impl super::behaviour::Conversion for PaymentMethodSession {
-    type DstType = diesel_models::payment_methods_session::PaymentMethodSession;
-    type NewDstType = diesel_models::payment_methods_session::PaymentMethodSession;
-    async fn convert(self) -> CustomResult<Self::DstType, ValidationError> {
-        Ok(Self::DstType {
-            id: self.id,
-            customer_id: self.customer_id,
-            billing: self.billing.map(|val| val.into()),
-            psp_tokenization: self.psp_tokenization,
-            network_tokenization: self.network_tokenization,
-            tokenization_data: self.tokenization_data,
-            expires_at: self.expires_at,
-            associated_payment_methods: self.associated_payment_methods,
-            associated_payment: self.associated_payment,
-            return_url: self.return_url,
-            associated_token_id: self.associated_token_id,
-            storage_type: self.storage_type,
-            keep_alive: self.keep_alive,
-        })
-    }
-
-    async fn convert_back(
-        state: &keymanager::KeyManagerState,
-        storage_model: Self::DstType,
-        key: &Secret<Vec<u8>>,
-        key_manager_identifier: keymanager::Identifier,
-    ) -> CustomResult<Self, ValidationError>
-    where
-        Self: Sized,
-    {
-        use common_utils::ext_traits::ValueExt;
-
-        async {
-            let decrypted_data = crypto_operation(
-                state,
-                type_name!(Self::DstType),
-                CryptoOperation::BatchDecrypt(EncryptedPaymentMethodSession::to_encryptable(
-                    EncryptedPaymentMethodSession {
-                        billing: storage_model.billing,
-                    },
-                )),
-                key_manager_identifier,
-                key.peek(),
-            )
-            .await
-            .and_then(|val| val.try_into_batchoperation())?;
-
-            let data = EncryptedPaymentMethodSession::from_encryptable(decrypted_data)
-                .change_context(common_utils::errors::CryptoError::DecodingFailed)
-                .attach_printable("Invalid batch operation data")?;
-
-            let billing = data
-                .billing
-                .map(|billing| {
-                    billing.deserialize_inner_value(|value| value.parse_value("Address"))
-                })
-                .transpose()
-                .change_context(common_utils::errors::CryptoError::DecodingFailed)
-                .attach_printable("Error while deserializing Address")?;
-
-            Ok::<Self, error_stack::Report<common_utils::errors::CryptoError>>(Self {
-                id: storage_model.id,
-                customer_id: storage_model.customer_id,
-                billing,
-                psp_tokenization: storage_model.psp_tokenization,
-                network_tokenization: storage_model.network_tokenization,
-                tokenization_data: storage_model.tokenization_data,
-                expires_at: storage_model.expires_at,
-                associated_payment_methods: storage_model.associated_payment_methods,
-                associated_payment: storage_model.associated_payment,
-                return_url: storage_model.return_url,
-                associated_token_id: storage_model.associated_token_id,
-                storage_type: storage_model.storage_type,
-                keep_alive: storage_model.keep_alive,
-            })
-        }
-        .await
-        .change_context(ValidationError::InvalidValue {
-            message: "Failed while decrypting payment method data".to_string(),
-        })
-    }
-
-    async fn construct_new(self) -> CustomResult<Self::NewDstType, ValidationError> {
-        Ok(Self::NewDstType {
-            id: self.id,
-            customer_id: self.customer_id,
-            billing: self.billing.map(|val| val.into()),
-            psp_tokenization: self.psp_tokenization,
-            network_tokenization: self.network_tokenization,
-            tokenization_data: self.tokenization_data,
-            expires_at: self.expires_at,
-            associated_payment_methods: self.associated_payment_methods,
-            associated_payment: self.associated_payment,
-            return_url: self.return_url,
-            associated_token_id: self.associated_token_id,
-            storage_type: self.storage_type,
-            keep_alive: self.keep_alive,
-        })
-    }
 }
 
 #[async_trait::async_trait]
