+++ crates/router/src/db/payment_method_session.rs
@@ -42,9 +42,11 @@ impl PaymentMethodsSessionInterface for crate::services::Store {}
 #[cfg(feature = "v2")]
 mod storage {
     use error_stack::ResultExt;
-    use hyperswitch_domain_models::behaviour::{Conversion, ReverseConversion};
     use router_env::{instrument, tracing};
-    use storage_impl::redis::kv_store::RedisConnInterface;
+    use storage_impl::{
+        behaviour::{Conversion, ReverseConversion},
+        redis::kv_store::RedisConnInterface,
+    };
 
     use super::PaymentMethodsSessionInterface;
     use crate::{
