+++ crates/storage_impl/src/customers.rs
@@ -1,16 +1,15 @@
 use common_utils::{id_type, pii};
-use diesel_models::{customers, kv};
+use diesel_models::kv;
 use error_stack::ResultExt;
 use futures::future::try_join_all;
 use hyperswitch_domain_models::{
-    behaviour::{Conversion, ReverseConversion},
-    customer as domain,
-    merchant_key_store::MerchantKeyStore,
+    customer as domain, merchant_key_store::MerchantKeyStore, type_encryption::AsyncLift,
 };
 use hyperswitch_masking::PeekInterface;
 use router_env::{instrument, tracing};
 
 use crate::{
+    behaviour::{Conversion, ForeignFrom, ForeignInto, ReverseConversion},
     diesel_error_to_data_error,
     errors::StorageError,
     kv_router_store,
@@ -20,7 +19,7 @@ use crate::{
     CustomResult, DatabaseStore, MockDb, RouterStore,
 };
 
-impl KvStorePartition for customers::Customer {}
+impl KvStorePartition for diesel_models::Customer {}
 
 #[cfg(feature = "v2")]
 mod label {
@@ -66,7 +65,7 @@ impl<T: DatabaseStore> domain::CustomerInterface for kv_router_store::KVRouterSt
             .find_optional_resource_by_id(
                 key_store,
                 storage_scheme,
-                customers::Customer::find_optional_by_customer_id_merchant_id(
+                diesel_models::Customer::find_optional_by_customer_id_merchant_id(
                     &conn,
                     customer_id,
                     merchant_id,
@@ -101,7 +100,7 @@ impl<T: DatabaseStore> domain::CustomerInterface for kv_router_store::KVRouterSt
         self.find_optional_resource_by_id(
             key_store,
             storage_scheme,
-            customers::Customer::find_optional_by_customer_id_merchant_id(
+            diesel_models::Customer::find_optional_by_customer_id_merchant_id(
                 &conn,
                 customer_id,
                 merchant_id,
@@ -130,7 +129,7 @@ impl<T: DatabaseStore> domain::CustomerInterface for kv_router_store::KVRouterSt
             .find_optional_resource_by_id(
                 key_store,
                 storage_scheme,
-                customers::Customer::find_optional_by_merchant_id_merchant_reference_id(
+                diesel_models::Customer::find_optional_by_merchant_id_merchant_reference_id(
                     &conn,
                     merchant_reference_id,
                     merchant_id,
@@ -166,8 +165,9 @@ impl<T: DatabaseStore> domain::CustomerInterface for kv_router_store::KVRouterSt
         let customer = Conversion::convert(customer)
             .await
             .change_context(StorageError::EncryptionError)?;
-        let updated_customer = diesel_models::CustomerUpdateInternal::from(customer_update.clone())
-            .apply_changeset(customer.clone());
+        let updated_customer =
+            diesel_models::CustomerUpdateInternal::foreign_from(customer_update.clone())
+                .apply_changeset(customer.clone());
         let key = PartitionKey::MerchantIdCustomerId {
             merchant_id: &merchant_id,
             customer_id: &customer_id,
@@ -176,17 +176,17 @@ impl<T: DatabaseStore> domain::CustomerInterface for kv_router_store::KVRouterSt
         self.update_resource(
             key_store,
             storage_scheme,
-            customers::Customer::update_by_customer_id_merchant_id(
+            diesel_models::Customer::update_by_customer_id_merchant_id(
                 &conn,
                 customer_id.clone(),
                 merchant_id.clone(),
-                customer_update.clone().into(),
+                customer_update.clone().foreign_into(),
             ),
             updated_customer,
             kv_router_store::UpdateResourceParams {
                 updateable: kv::Updateable::CustomerUpdate(Box::new(kv::CustomerUpdateMems {
                     orig: customer.clone(),
-                    update_data: customer_update.clone().into(),
+                    update_data: customer_update.clone().foreign_into(),
                 })),
                 operation: Op::Update(key.clone(), &field, customer.updated_by.as_deref()),
             },
@@ -208,7 +208,7 @@ impl<T: DatabaseStore> domain::CustomerInterface for kv_router_store::KVRouterSt
             .find_resource_by_id(
                 key_store,
                 storage_scheme,
-                customers::Customer::find_by_merchant_reference_id_merchant_id(
+                diesel_models::Customer::find_by_merchant_reference_id_merchant_id(
                     &conn,
                     merchant_reference_id,
                     merchant_id,
@@ -243,7 +243,7 @@ impl<T: DatabaseStore> domain::CustomerInterface for kv_router_store::KVRouterSt
             .find_resource_by_id(
                 key_store,
                 storage_scheme,
-                customers::Customer::find_by_customer_id_merchant_id(
+                diesel_models::Customer::find_by_customer_id_merchant_id(
                     &conn,
                     customer_id,
                     merchant_id,
@@ -307,7 +307,7 @@ impl<T: DatabaseStore> domain::CustomerInterface for kv_router_store::KVRouterSt
             .await
             .change_context(StorageError::EncryptionError)?;
 
-        let decided_storage_scheme = Box::pin(decide_storage_scheme::<_, customers::Customer>(
+        let decided_storage_scheme = Box::pin(decide_storage_scheme::<_, diesel_models::Customer>(
             self,
             storage_scheme,
             Op::Insert,
@@ -357,7 +357,7 @@ impl<T: DatabaseStore> domain::CustomerInterface for kv_router_store::KVRouterSt
             .construct_new()
             .await
             .change_context(StorageError::EncryptionError)?;
-        let storage_scheme = Box::pin(decide_storage_scheme::<_, customers::Customer>(
+        let storage_scheme = Box::pin(decide_storage_scheme::<_, diesel_models::Customer>(
             self,
             storage_scheme,
             Op::Insert,
@@ -406,7 +406,7 @@ impl<T: DatabaseStore> domain::CustomerInterface for kv_router_store::KVRouterSt
             .find_resource_by_id(
                 key_store,
                 storage_scheme,
-                customers::Customer::find_by_global_id(&conn, id),
+                diesel_models::Customer::find_by_global_id(&conn, id),
                 kv_router_store::FindResourceBy::Id(
                     format!("cust_{}", id.get_string_repr()),
                     PartitionKey::GlobalId {
@@ -437,7 +437,7 @@ impl<T: DatabaseStore> domain::CustomerInterface for kv_router_store::KVRouterSt
             .find_resource_by_id(
                 key_store,
                 storage_scheme,
-                customers::Customer::find_by_global_id_merchant_id(&conn, id, merchant_id),
+                diesel_models::Customer::find_by_global_id_merchant_id(&conn, id, merchant_id),
                 kv_router_store::FindResourceBy::Id(
                     format!("cust_{}", id.get_string_repr()),
                     PartitionKey::GlobalId {
@@ -472,8 +472,11 @@ impl<T: DatabaseStore> domain::CustomerInterface for kv_router_store::KVRouterSt
         let customer = Conversion::convert(customer)
             .await
             .change_context(StorageError::EncryptionError)?;
-        let database_call =
-            customers::Customer::update_by_id(&conn, id.clone(), customer_update.clone().into());
+        let database_call = diesel_models::Customer::update_by_id(
+            &conn,
+            id.clone(),
+            customer_update.clone().foreign_into(),
+        );
         let key = PartitionKey::GlobalId {
             id: id.get_string_repr(),
         };
@@ -482,12 +485,12 @@ impl<T: DatabaseStore> domain::CustomerInterface for kv_router_store::KVRouterSt
             key_store,
             storage_scheme,
             database_call,
-            diesel_models::CustomerUpdateInternal::from(customer_update.clone())
+            diesel_models::CustomerUpdateInternal::foreign_from(customer_update.clone())
                 .apply_changeset(customer.clone()),
             kv_router_store::UpdateResourceParams {
                 updateable: kv::Updateable::CustomerUpdate(Box::new(kv::CustomerUpdateMems {
                     orig: customer.clone(),
-                    update_data: customer_update.into(),
+                    update_data: customer_update.foreign_into(),
                 })),
                 operation: Op::Update(key.clone(), &field, customer.updated_by.as_deref()),
             },
@@ -510,9 +513,9 @@ impl<T: DatabaseStore> domain::CustomerInterface for RouterStore<T> {
     ) -> CustomResult<Option<domain::Customer>, StorageError> {
         let conn = pg_connection_read(self).await?;
         let maybe_customer: Option<domain::Customer> = self
-            .find_optional_resource(
+            .find_optional_resource_new(
                 key_store,
-                customers::Customer::find_optional_by_customer_id_merchant_id(
+                diesel_models::Customer::find_optional_by_customer_id_merchant_id(
                     &conn,
                     customer_id,
                     merchant_id,
@@ -541,9 +544,9 @@ impl<T: DatabaseStore> domain::CustomerInterface for RouterStore<T> {
         _storage_scheme: MerchantStorageScheme,
     ) -> CustomResult<Option<domain::Customer>, StorageError> {
         let conn = pg_connection_read(self).await?;
-        self.find_optional_resource(
+        self.find_optional_resource_new(
             key_store,
-            customers::Customer::find_optional_by_customer_id_merchant_id(
+            diesel_models::Customer::find_optional_by_customer_id_merchant_id(
                 &conn,
                 customer_id,
                 merchant_id,
@@ -563,9 +566,9 @@ impl<T: DatabaseStore> domain::CustomerInterface for RouterStore<T> {
     ) -> CustomResult<Option<domain::Customer>, StorageError> {
         let conn = pg_connection_read(self).await?;
         let maybe_customer: Option<domain::Customer> = self
-            .find_optional_resource(
+            .find_optional_resource_new(
                 key_store,
-                customers::Customer::find_optional_by_merchant_id_merchant_reference_id(
+                diesel_models::Customer::find_optional_by_merchant_id_merchant_reference_id(
                     &conn,
                     customer_id,
                     merchant_id,
@@ -596,13 +599,13 @@ impl<T: DatabaseStore> domain::CustomerInterface for RouterStore<T> {
         _storage_scheme: MerchantStorageScheme,
     ) -> CustomResult<domain::Customer, StorageError> {
         let conn = pg_connection_write(self).await?;
-        self.call_database(
+        self.call_database_new(
             key_store,
-            customers::Customer::update_by_customer_id_merchant_id(
+            diesel_models::Customer::update_by_customer_id_merchant_id(
                 &conn,
                 customer_id,
                 merchant_id.clone(),
-                customer_update.into(),
+                customer_update.foreign_into(),
             ),
         )
         .await
@@ -619,9 +622,9 @@ impl<T: DatabaseStore> domain::CustomerInterface for RouterStore<T> {
     ) -> CustomResult<domain::Customer, StorageError> {
         let conn = pg_connection_read(self).await?;
         let customer: domain::Customer = self
-            .call_database(
+            .call_database_new(
                 key_store,
-                customers::Customer::find_by_customer_id_merchant_id(
+                diesel_models::Customer::find_by_customer_id_merchant_id(
                     &conn,
                     customer_id,
                     merchant_id,
@@ -645,9 +648,9 @@ impl<T: DatabaseStore> domain::CustomerInterface for RouterStore<T> {
     ) -> CustomResult<domain::Customer, StorageError> {
         let conn = pg_connection_read(self).await?;
         let customer: domain::Customer = self
-            .call_database(
+            .call_database_new(
                 key_store,
-                customers::Customer::find_by_merchant_reference_id_merchant_id(
+                diesel_models::Customer::find_by_merchant_reference_id_merchant_id(
                     &conn,
                     merchant_reference_id,
                     merchant_id,
@@ -670,9 +673,9 @@ impl<T: DatabaseStore> domain::CustomerInterface for RouterStore<T> {
         let conn = pg_connection_read(self).await?;
         let customer_list_constraints =
             diesel_models::query::customers::CustomerListConstraints::from(constraints);
-        self.find_resources(
+        self.find_resources_new(
             key_store,
-            customers::Customer::list_customers_by_merchant_id_and_constraints(
+            diesel_models::Customer::list_customers_by_merchant_id_and_constraints(
                 &conn,
                 merchant_id,
                 customer_list_constraints,
@@ -698,25 +701,26 @@ impl<T: DatabaseStore> domain::CustomerInterface for RouterStore<T> {
             time_range: customer_list_constraints.time_range,
         };
         let customers = self
-            .find_resources(
+            .find_resources_new(
                 key_store,
-                customers::Customer::list_customers_by_merchant_id_and_constraints(
+                diesel_models::Customer::list_customers_by_merchant_id_and_constraints(
                     &conn,
                     merchant_id,
                     customers_constraints,
                 ),
             )
             .await?;
-        let total_count = customers::Customer::get_customer_count_by_merchant_id_and_constraints(
-            &conn,
-            merchant_id,
-            customer_list_constraints,
-        )
-        .await
-        .map_err(|error| {
-            let new_err = diesel_error_to_data_error(*error.current_context());
-            error.change_context(new_err)
-        })?;
+        let total_count =
+            diesel_models::Customer::get_customer_count_by_merchant_id_and_constraints(
+                &conn,
+                merchant_id,
+                customer_list_constraints,
+            )
+            .await
+            .map_err(|error| {
+                let new_err = diesel_error_to_data_error(*error.current_context());
+                error.change_context(new_err)
+            })?;
         Ok((customers, total_count))
     }
 
@@ -732,7 +736,7 @@ impl<T: DatabaseStore> domain::CustomerInterface for RouterStore<T> {
             .construct_new()
             .await
             .change_context(StorageError::EncryptionError)?;
-        self.call_database(key_store, customer_new.insert(&conn))
+        self.call_database_new(key_store, customer_new.insert(&conn))
             .await
     }
 
@@ -744,7 +748,7 @@ impl<T: DatabaseStore> domain::CustomerInterface for RouterStore<T> {
         merchant_id: &id_type::MerchantId,
     ) -> CustomResult<bool, StorageError> {
         let conn = pg_connection_write(self).await?;
-        customers::Customer::delete_by_customer_id_merchant_id(&conn, customer_id, merchant_id)
+        diesel_models::Customer::delete_by_customer_id_merchant_id(&conn, customer_id, merchant_id)
             .await
             .map_err(|error| {
                 let new_err = diesel_error_to_data_error(*error.current_context());
@@ -763,9 +767,13 @@ impl<T: DatabaseStore> domain::CustomerInterface for RouterStore<T> {
         _storage_scheme: MerchantStorageScheme,
     ) -> CustomResult<domain::Customer, StorageError> {
         let conn = pg_connection_write(self).await?;
-        self.call_database(
+        self.call_database_new(
             key_store,
-            customers::Customer::update_by_id(&conn, id.clone(), customer_update.into()),
+            diesel_models::Customer::update_by_id(
+                &conn,
+                id.clone(),
+                customer_update.foreign_into(),
+            ),
         )
         .await
     }
@@ -780,7 +788,10 @@ impl<T: DatabaseStore> domain::CustomerInterface for RouterStore<T> {
     ) -> CustomResult<domain::Customer, StorageError> {
         let conn = pg_connection_read(self).await?;
         let customer: domain::Customer = self
-            .call_database(key_store, customers::Customer::find_by_global_id(&conn, id))
+            .call_database_new(
+                key_store,
+                diesel_models::Customer::find_by_global_id(&conn, id),
+            )
             .await?;
         match customer.name {
             Some(ref name) if name.peek() == pii::REDACTED => Err(StorageError::CustomerRedacted)?,
@@ -799,9 +810,9 @@ impl<T: DatabaseStore> domain::CustomerInterface for RouterStore<T> {
     ) -> CustomResult<domain::Customer, StorageError> {
         let conn = pg_connection_read(self).await?;
         let customer: domain::Customer = self
-            .call_database(
+            .call_database_new(
                 key_store,
-                customers::Customer::find_by_global_id_merchant_id(&conn, id, merchant_id),
+                diesel_models::Customer::find_by_global_id_merchant_id(&conn, id, merchant_id),
             )
             .await?;
         match customer.name {
@@ -823,7 +834,7 @@ impl domain::CustomerInterface for MockDb {
         _storage_scheme: MerchantStorageScheme,
     ) -> CustomResult<Option<domain::Customer>, StorageError> {
         let customers = self.customers.lock().await;
-        self.find_resource(key_store, customers, |customer| {
+        self.find_resource_new(key_store, customers, |customer| {
             customer.customer_id == *customer_id && &customer.merchant_id == merchant_id
         })
         .await
@@ -838,7 +849,7 @@ impl domain::CustomerInterface for MockDb {
         _storage_scheme: MerchantStorageScheme,
     ) -> CustomResult<Option<domain::Customer>, StorageError> {
         let customers = self.customers.lock().await;
-        self.find_resource(key_store, customers, |customer| {
+        self.find_resource_new(key_store, customers, |customer| {
             customer.customer_id == *customer_id && &customer.merchant_id == merchant_id
         })
         .await
@@ -1032,5 +1043,517 @@ impl domain::CustomerInterface for MockDb {
     ) -> CustomResult<domain::Customer, StorageError> {
         // [#172]: Implement function for `MockDb`
         Err(StorageError::MockDbError)?
+    }
+}
+
+#[cfg(feature = "v2")]
+use common_enums::DeleteStatus;
+use common_utils::{
+    crypto::Encryptable,
+    date_time,
+    encryption::Encryption,
+    errors::ValidationError,
+    types::{
+        keymanager::{self, KeyManagerState, ToEncryptable},
+        CreatedBy,
+    },
+};
+use hyperswitch_domain_models::type_encryption;
+use hyperswitch_masking::{Secret, SwitchStrategy};
+
+#[cfg(feature = "v1")]
+#[async_trait::async_trait]
+impl Conversion for domain::Customer {
+    type DstType = diesel_models::Customer;
+    type NewDstType = diesel_models::CustomerNew;
+    async fn convert(self) -> CustomResult<Self::DstType, ValidationError> {
+        Ok(diesel_models::Customer {
+            customer_id: self.customer_id.clone(),
+            merchant_id: self.merchant_id,
+            name: self.name.map(Encryption::from),
+            email: self.email.map(Encryption::from),
+            phone: self.phone.map(Encryption::from),
+            phone_country_code: self.phone_country_code,
+            description: self.description,
+            created_at: self.created_at,
+            metadata: self.metadata,
+            modified_at: self.modified_at,
+            connector_customer: self.connector_customer,
+            address_id: self.address_id,
+            default_payment_method_id: self.default_payment_method_id,
+            updated_by: self.updated_by,
+            version: self.version,
+            tax_registration_id: self.tax_registration_id.map(Encryption::from),
+            document_details: self.document_details.map(Encryption::from),
+            created_by: self.created_by.map(|created_by| created_by.to_string()),
+            last_modified_by: self
+                .last_modified_by
+                .map(|last_modified_by| last_modified_by.to_string()),
+            id: Some(self.customer_id),
+        })
+    }
+
+    async fn convert_back(
+        state: &KeyManagerState,
+        item: Self::DstType,
+        key: &Secret<Vec<u8>>,
+        _key_store_ref_id: keymanager::Identifier,
+    ) -> CustomResult<Self, ValidationError>
+    where
+        Self: Sized,
+    {
+        let decrypted = type_encryption::crypto_operation(
+            state,
+            common_utils::type_name!(Self::DstType),
+            type_encryption::CryptoOperation::BatchDecrypt(
+                domain::EncryptedCustomer::to_encryptable(domain::EncryptedCustomer {
+                    name: item.name.clone(),
+                    phone: item.phone.clone(),
+                    email: item.email.clone(),
+                    tax_registration_id: item.tax_registration_id.clone(),
+                }),
+            ),
+            keymanager::Identifier::Merchant(item.merchant_id.clone()),
+            key.peek(),
+        )
+        .await
+        .and_then(|val| val.try_into_batchoperation())
+        .change_context(ValidationError::InvalidValue {
+            message: "Failed while decrypting customer data".to_string(),
+        })?;
+        let encryptable_customer = domain::EncryptedCustomer::from_encryptable(decrypted)
+            .change_context(ValidationError::InvalidValue {
+                message: "Failed while decrypting customer data".to_string(),
+            })?;
+        let document_details = item
+            .document_details
+            .async_lift(|inner| async {
+                type_encryption::crypto_operation(
+                    state,
+                    common_utils::type_name!(Self),
+                    type_encryption::CryptoOperation::DecryptOptional(inner),
+                    keymanager::Identifier::Merchant(item.merchant_id.clone()),
+                    key.peek(),
+                )
+                .await
+                .and_then(|val| val.try_into_optionaloperation())
+            })
+            .await
+            .change_context(ValidationError::InvalidValue {
+                message: "Failed to decrypt document details".to_string(),
+            })?;
+
+        Ok(Self {
+            customer_id: item.customer_id,
+            merchant_id: item.merchant_id,
+            name: encryptable_customer.name,
+            email: encryptable_customer.email.map(|email| {
+                let encryptable: Encryptable<Secret<String, pii::EmailStrategy>> = Encryptable::new(
+                    email.clone().into_inner().switch_strategy(),
+                    email.into_encrypted(),
+                );
+                encryptable
+            }),
+            phone: encryptable_customer.phone,
+            phone_country_code: item.phone_country_code,
+            description: item.description,
+            created_at: item.created_at,
+            metadata: item.metadata,
+            modified_at: item.modified_at,
+            connector_customer: item.connector_customer,
+            address_id: item.address_id,
+            default_payment_method_id: item.default_payment_method_id,
+            updated_by: item.updated_by,
+            version: item.version,
+            tax_registration_id: encryptable_customer.tax_registration_id,
+            document_details,
+            created_by: item
+                .created_by
+                .and_then(|created_by| created_by.parse::<CreatedBy>().ok()),
+            last_modified_by: item
+                .last_modified_by
+                .and_then(|last_modified_by| last_modified_by.parse::<CreatedBy>().ok()),
+        })
+    }
+
+    async fn construct_new(self) -> CustomResult<Self::NewDstType, ValidationError> {
+        let now = date_time::now();
+        Ok(diesel_models::CustomerNew {
+            id: Some(self.customer_id.clone()),
+            customer_id: self.customer_id,
+            merchant_id: self.merchant_id,
+            name: self.name.map(Encryption::from),
+            email: self.email.map(Encryption::from),
+            phone: self.phone.map(Encryption::from),
+            description: self.description,
+            phone_country_code: self.phone_country_code,
+            metadata: self.metadata,
+            created_at: now,
+            modified_at: now,
+            connector_customer: self.connector_customer,
+            address_id: self.address_id,
+            updated_by: self.updated_by,
+            version: self.version,
+            tax_registration_id: self.tax_registration_id.map(Encryption::from),
+            document_details: self.document_details.map(Encryption::from),
+            created_by: self
+                .created_by
+                .as_ref()
+                .map(|created_by| created_by.to_string()),
+            last_modified_by: self.created_by.map(|created_by| created_by.to_string()), // Same as created_by on creation
+        })
+    }
+}
+
+#[cfg(feature = "v1")]
+impl ForeignFrom<domain::CustomerUpdate> for diesel_models::CustomerUpdateInternal {
+    fn foreign_from(customer_update: domain::CustomerUpdate) -> Self {
+        match customer_update {
+            domain::CustomerUpdate::Update {
+                name,
+                email,
+                phone,
+                description,
+                phone_country_code,
+                metadata,
+                connector_customer,
+                address_id,
+                tax_registration_id,
+                document_details,
+                last_modified_by,
+            } => Self {
+                name: name.map(Encryption::from),
+                email: email.map(Encryption::from),
+                phone: phone.map(Encryption::from),
+                description,
+                phone_country_code,
+                metadata: *metadata,
+                connector_customer: *connector_customer,
+                modified_at: date_time::now(),
+                address_id,
+                default_payment_method_id: None,
+                updated_by: None,
+                tax_registration_id: tax_registration_id.map(Encryption::from),
+                document_details: document_details.map(Encryption::from),
+                last_modified_by,
+            },
+            domain::CustomerUpdate::ConnectorCustomer {
+                connector_customer,
+                last_modified_by,
+            } => Self {
+                connector_customer,
+                modified_at: date_time::now(),
+                name: None,
+                email: None,
+                phone: None,
+                description: None,
+                phone_country_code: None,
+                metadata: None,
+                default_payment_method_id: None,
+                updated_by: None,
+                address_id: None,
+                tax_registration_id: None,
+                document_details: None,
+                last_modified_by,
+            },
+            domain::CustomerUpdate::UpdateDefaultPaymentMethod {
+                default_payment_method_id,
+                last_modified_by,
+            } => Self {
+                default_payment_method_id,
+                modified_at: date_time::now(),
+                name: None,
+                email: None,
+                phone: None,
+                description: None,
+                phone_country_code: None,
+                metadata: None,
+                connector_customer: None,
+                updated_by: None,
+                address_id: None,
+                tax_registration_id: None,
+                document_details: None,
+                last_modified_by,
+            },
+        }
+    }
+}
+
+#[cfg(feature = "v2")]
+#[async_trait::async_trait]
+impl Conversion for domain::Customer {
+    type DstType = diesel_models::Customer;
+    type NewDstType = diesel_models::CustomerNew;
+    async fn convert(self) -> CustomResult<Self::DstType, ValidationError> {
+        Ok(diesel_models::Customer {
+            id: self.id.clone(),
+            customer_id: Some(self.id),
+            merchant_reference_id: self.merchant_reference_id,
+            merchant_id: self.merchant_id,
+            name: self.name.map(Encryption::from),
+            email: self.email.map(Encryption::from),
+            phone: self.phone.map(Encryption::from),
+            phone_country_code: self.phone_country_code,
+            description: self.description,
+            created_at: self.created_at,
+            metadata: self.metadata,
+            modified_at: self.modified_at,
+            connector_customer: self.connector_customer,
+            default_payment_method_id: self.default_payment_method_id,
+            updated_by: self.updated_by,
+            default_billing_address: self.default_billing_address.map(Encryption::from),
+            default_shipping_address: self.default_shipping_address.map(Encryption::from),
+            version: self.version,
+            status: self.status,
+            tax_registration_id: self.tax_registration_id.map(Encryption::from),
+            document_details: self.document_details.map(Encryption::from),
+            created_by: self.created_by.map(|created_by| created_by.to_string()),
+            last_modified_by: self
+                .last_modified_by
+                .map(|last_modified_by| last_modified_by.to_string()),
+        })
+    }
+
+    async fn convert_back(
+        state: &KeyManagerState,
+        item: Self::DstType,
+        key: &Secret<Vec<u8>>,
+        _key_store_ref_id: keymanager::Identifier,
+    ) -> CustomResult<Self, ValidationError>
+    where
+        Self: Sized,
+    {
+        let decrypted = type_encryption::crypto_operation(
+            state,
+            common_utils::type_name!(Self::DstType),
+            type_encryption::CryptoOperation::BatchDecrypt(
+                domain::EncryptedCustomer::to_encryptable(domain::EncryptedCustomer {
+                    name: item.name.clone(),
+                    phone: item.phone.clone(),
+                    email: item.email.clone(),
+                    tax_registration_id: item.tax_registration_id.clone(),
+                }),
+            ),
+            keymanager::Identifier::Merchant(item.merchant_id.clone()),
+            key.peek(),
+        )
+        .await
+        .and_then(|val| val.try_into_batchoperation())
+        .change_context(ValidationError::InvalidValue {
+            message: "Failed while decrypting customer data".to_string(),
+        })?;
+        let encryptable_customer = domain::EncryptedCustomer::from_encryptable(decrypted)
+            .change_context(ValidationError::InvalidValue {
+                message: "Failed while decrypting customer data".to_string(),
+            })?;
+
+        let default_billing_address = item
+            .default_billing_address
+            .async_lift(|inner| async {
+                type_encryption::crypto_operation(
+                    state,
+                    common_utils::type_name!(Self),
+                    type_encryption::CryptoOperation::DecryptOptional(inner),
+                    keymanager::Identifier::Merchant(item.merchant_id.clone()),
+                    key.peek(),
+                )
+                .await
+                .and_then(|val| val.try_into_optionaloperation())
+            })
+            .await
+            .change_context(ValidationError::InvalidValue {
+                message: "Failed to decrypt default billing address".to_string(),
+            })?;
+
+        let default_shipping_address = item
+            .default_shipping_address
+            .async_lift(|inner| async {
+                type_encryption::crypto_operation(
+                    state,
+                    common_utils::type_name!(Self),
+                    type_encryption::CryptoOperation::DecryptOptional(inner),
+                    keymanager::Identifier::Merchant(item.merchant_id.clone()),
+                    key.peek(),
+                )
+                .await
+                .and_then(|val| val.try_into_optionaloperation())
+            })
+            .await
+            .change_context(ValidationError::InvalidValue {
+                message: "Failed to decrypt default shipping address".to_string(),
+            })?;
+
+        let document_details = item
+            .document_details
+            .async_lift(|inner| async {
+                type_encryption::crypto_operation(
+                    state,
+                    common_utils::type_name!(Self),
+                    type_encryption::CryptoOperation::DecryptOptional(inner),
+                    keymanager::Identifier::Merchant(item.merchant_id.clone()),
+                    key.peek(),
+                )
+                .await
+                .and_then(|val| val.try_into_optionaloperation())
+            })
+            .await
+            .change_context(ValidationError::InvalidValue {
+                message: "Failed to decrypt document details".to_string(),
+            })?;
+
+        Ok(Self {
+            id: item.id,
+            merchant_reference_id: item.merchant_reference_id,
+            merchant_id: item.merchant_id,
+            name: encryptable_customer.name,
+            email: encryptable_customer.email.map(|email| {
+                let encryptable: Encryptable<Secret<String, pii::EmailStrategy>> = Encryptable::new(
+                    email.clone().into_inner().switch_strategy(),
+                    email.into_encrypted(),
+                );
+                encryptable
+            }),
+            phone: encryptable_customer.phone,
+            phone_country_code: item.phone_country_code,
+            description: item.description,
+            created_at: item.created_at,
+            metadata: item.metadata,
+            modified_at: item.modified_at,
+            connector_customer: item.connector_customer,
+            default_payment_method_id: item.default_payment_method_id,
+            updated_by: item.updated_by,
+            default_billing_address,
+            default_shipping_address,
+            version: item.version,
+            status: item.status,
+            tax_registration_id: encryptable_customer.tax_registration_id,
+            document_details,
+            created_by: item
+                .created_by
+                .and_then(|created_by| created_by.parse::<CreatedBy>().ok()),
+            last_modified_by: item
+                .last_modified_by
+                .and_then(|last_modified_by| last_modified_by.parse::<CreatedBy>().ok()),
+        })
+    }
+
+    async fn construct_new(self) -> CustomResult<Self::NewDstType, ValidationError> {
+        let now = date_time::now();
+        Ok(diesel_models::customers::CustomerNew {
+            id: self.id.clone(),
+            merchant_reference_id: self.merchant_reference_id,
+            merchant_id: self.merchant_id,
+            name: self.name.map(Encryption::from),
+            email: self.email.map(Encryption::from),
+            phone: self.phone.map(Encryption::from),
+            description: self.description,
+            phone_country_code: self.phone_country_code,
+            metadata: self.metadata,
+            default_payment_method_id: None,
+            created_at: now,
+            modified_at: now,
+            connector_customer: self.connector_customer,
+            updated_by: self.updated_by,
+            default_billing_address: self.default_billing_address.map(Encryption::from),
+            default_shipping_address: self.default_shipping_address.map(Encryption::from),
+            // TODO: Flag this in review
+            // version: common_types::consts::API_VERSION,
+            version: self.version,
+            status: self.status,
+            tax_registration_id: self.tax_registration_id.map(Encryption::from),
+            document_details: self.document_details.map(Encryption::from),
+            created_by: self
+                .created_by
+                .as_ref()
+                .map(|created_by| created_by.to_string()),
+            last_modified_by: self.created_by.map(|created_by| created_by.to_string()), // Same as created_by on creation
+            customer_id: Some(self.id),
+        })
+    }
+}
+
+#[cfg(feature = "v2")]
+impl ForeignFrom<domain::CustomerUpdate> for diesel_models::CustomerUpdateInternal {
+    fn foreign_from(customer_update: domain::CustomerUpdate) -> Self {
+        match customer_update {
+            domain::CustomerUpdate::Update(update) => {
+                let domain::CustomerGeneralUpdate {
+                    name,
+                    email,
+                    phone,
+                    description,
+                    phone_country_code,
+                    metadata,
+                    connector_customer,
+                    default_billing_address,
+                    default_shipping_address,
+                    default_payment_method_id,
+                    status,
+                    tax_registration_id,
+                    document_details,
+                    last_modified_by,
+                } = *update;
+                Self {
+                    name: name.map(Encryption::from),
+                    email: email.map(Encryption::from),
+                    phone: phone.map(Encryption::from),
+                    description,
+                    phone_country_code,
+                    metadata,
+                    connector_customer: *connector_customer,
+                    modified_at: date_time::now(),
+                    default_billing_address: default_billing_address.map(Encryption::from),
+                    default_shipping_address: default_shipping_address.map(Encryption::from),
+                    default_payment_method_id,
+                    updated_by: None,
+                    status,
+                    tax_registration_id: tax_registration_id.map(Encryption::from),
+                    document_details: document_details.map(Encryption::from),
+                    last_modified_by,
+                }
+            }
+            domain::CustomerUpdate::ConnectorCustomer {
+                connector_customer,
+                last_modified_by,
+            } => Self {
+                connector_customer,
+                name: None,
+                email: None,
+                phone: None,
+                description: None,
+                phone_country_code: None,
+                metadata: None,
+                modified_at: date_time::now(),
+                default_payment_method_id: None,
+                updated_by: None,
+                default_billing_address: None,
+                default_shipping_address: None,
+                status: None,
+                tax_registration_id: None,
+                document_details: None,
+                last_modified_by,
+            },
+            domain::CustomerUpdate::UpdateDefaultPaymentMethod {
+                default_payment_method_id,
+                last_modified_by,
+            } => Self {
+                default_payment_method_id,
+                modified_at: date_time::now(),
+                name: None,
+                email: None,
+                phone: None,
+                description: None,
+                phone_country_code: None,
+                metadata: None,
+                connector_customer: None,
+                updated_by: None,
+                default_billing_address: None,
+                default_shipping_address: None,
+                status: None,
+                tax_registration_id: None,
+                document_details: None,
+                last_modified_by,
+            },
+        }
     }
 }
